Squashed commit of the following:
commit d0e6276f24afd22936498ab650c844dd05964522
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 17:38:46 2009 +0900

    [#339] 共有ファイルの新規作成、編集ができないケースがあったので

commit 535fd37b5b3107651acf9a8fc9ec52b7340a1bae
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 15:04:04 2009 +0900

    [#339] downloadの権限チェックを修正

commit 761fdb3dd3cb2380aee38c0681fa440d65f6d5b7
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 14:08:42 2009 +0900

    [#339] 不要になったので

commit 3b653db49e16ee585423d2e104ea91f2bf1e38e6
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 14:08:17 2009 +0900

    [#339] clear_download_historyの権限チェックを修正

commit d8128d3e40bc8d05266662f6dfb74801b2136d41
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 13:25:19 2009 +0900

    [#339] destroyを修正

commit dfae0fc65df73e48fbd0af73e8ffa23b4d7e5fcb
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 13:13:29 2009 +0900

    [#339] download_as_csvを修正

commit f4fb953521011d8c2c91874314e8a10dcd37a276
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 12:32:28 2009 +0900

    [#339] specの並べ替え

commit 666390a9ca1a740a6c4b20dd730ea67af0da4bae
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 12:30:36 2009 +0900

    [#339] destroyの権限チェックを修正

commit e73ccaef7fe2635fbd320bf48fb4e44ba6033c02
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 12:24:35 2009 +0900

    [#339] updateのspecを追加。不要な記述の除去

commit ac3cc371e816a31e4f7f55837753eb232ea9c4e0
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 12:22:03 2009 +0900

    [#339] createのspecを追加及び、createの権限チェックを修正

commit 818fc4ae63e59ca8bca344928fe43340f80453a2
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 10:15:47 2009 +0900

    [#339] newのspecを追加

commit 186b2237c5642f8edbcf627e9e7ac97af92a7791
Author: maedana <maeda.na@gmail.com>
Date:   Thu Jan 8 09:13:50 2009 +0900

    [#339] share_fileのvalidate時に権限チェックをするように

commit e0d74f91be67d38398b835c61ffeca5f42dbc9bb
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 18:07:17 2009 +0900

    [#339] readable?とupdatable?で使うユーザをあらかじめ設定できるように

commit 5803111e0675a2426bb91eaca5ff5baca8519008
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 17:48:07 2009 +0900

    [#339] updateのspecを追加

commit 24602efc907c71d0d9ad1b6d38368a31ad0d0936
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 16:27:51 2009 +0900

    [#339] editの権限チェックを置き換え

commit 7a1bf5b3bf3beb71466786b5d243e33a3c1794b8
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 15:47:03 2009 +0900

    [#339] グループ参加状態、グループ管理権限を判定するメソッドを追加

commit f1d5ecde5a6205a5a1a9f101595ede3278b7f695
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 14:53:00 2009 +0900

    [#339] 動いてなかったので

commit b705e0e8943f607035c11b999c84fe97e5d01a9d
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 14:45:45 2009 +0900

    [#339] リファクタ

commit 3a275d793ae898f0ee54c862a4c831f7fdfb4f16
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 14:42:01 2009 +0900

    [#339] 共有ファイルの更新(削除)権限を判定するメソッドを追加

commit 5765dea427216f00c72fedb3b8c52377f71ea67d
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 14:25:18 2009 +0900

    [#339] リファクタ

commit 761b09f3d02a524f99d86108e45ad4b8612c9626
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 14:19:45 2009 +0900

    [#339] グループがオーナーの場合の公開範囲指定メソッド追加等

commit 7bd89ef9e8544eb885aea7000b6b82d15843c183
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 11:46:41 2009 +0900

    [#339] リファクタ

commit f6825ba92db186a0d63ab035b329955054092711
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 11:43:57 2009 +0900

    [#339] ユーザがオーナーの場合の公開範囲判定メソッドを追加

commit e69a1d12eef12b7c3a16301bbde5b8af827e1ec5
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 10:49:57 2009 +0900

    [#339] 重複ロジックの排除

commit 080ef40be863c8c3b217f322c1a5163b7fa45f01
Author: maedana <maeda.na@gmail.com>
Date:   Wed Jan 7 09:50:35 2009 +0900

    [#339] リファクタ

commit 07611edebd2fb24d72c363c949b9ae79134c7635
Author: maedana <maeda.na@gmail.com>
Date:   Tue Jan 6 17:40:16 2009 +0900

    [#339] リファクタ(まだリファクタの余地あり)

commit e73e1353be9d1c5eb302ca42ef6b8d0ab1745987
Author: maedana <maeda.na@gmail.com>
Date:   Tue Jan 6 14:53:38 2009 +0900

    [#339] 共有ファイルの閲覧権限に関するspecの中身を記述して実装。(mockしてるメソッドはまだ未実装)

+++ app/controllers/share_file_controller.rb
@@ -20,8 +20,12 @@ class ShareFileController < ApplicationController
          :redirect_to => { :action => :index }
 
   def new
-    @share_file = ShareFile.new
-    @share_file.owner_symbol = params[:owner_symbol]
+    @share_file = ShareFile.new(:user_id => current_user.id, :owner_symbol => params[:owner_symbol])
+
+    unless @share_file.updatable?(current_user)
+      render_window_close
+      return
+    end
 
     params[:publication_type] = "public" if Symbol.split_symbol(@share_file.owner_symbol).first=="uid"
     params[:publication_type] ||= "private"
@@ -35,17 +39,14 @@ class ShareFileController < ApplicationController
   # post action
   def create
     @error_messages = []
-    @share_file = ShareFile.new(params[:share_file])
-    @share_file.user_id = session[:user_id]
-    @share_file.publication_type = params[:publication_type]
-    @share_file.publication_symbols_value = params[:publication_type]=='protected' ? params[:publication_symbols_value] : ""
-
-    # 所有者がマイユーザ OR マイグループ
-    unless login_user_symbols.include? @share_file.owner_symbol
-      @categories_hash = ShareFile.get_tags_hash(@share_file.owner_symbol)
-      redirect_to_with_deny_auth
+    unless params[:file]
+      render_window_close
       return
     end
+    @share_file = ShareFile.new(params[:share_file])
+    @share_file.user_id = current_user.id
+    @share_file.publication_type = params[:publication_type]
+    @share_file.publication_symbols_value = params[:publication_type]=='protected' ? params[:publication_symbols_value] : ""
 
     params[:file].each do |key, file|
       share_file = @share_file.clone
@@ -54,6 +55,7 @@ class ShareFileController < ApplicationController
         share_file.content_type = file.content_type.chomp
       end
       share_file.file = file
+      share_file.accessed_user = current_user
 
       if share_file.save
         target_symbols = analyze_param_publication_type
@@ -93,7 +95,7 @@ class ShareFileController < ApplicationController
       return
     end
 
-    unless authorize_to_save_share_file? @share_file
+    unless @share_file.updatable?(current_user)
       render_window_close
       return
     end
@@ -119,12 +121,7 @@ class ShareFileController < ApplicationController
     update_params = params[:share_file]
     update_params[:publication_type] = params[:publication_type]
     update_params[:publication_symbols_value] = params[:publication_type]=='protected' ? params[:publication_symbols_value] : ""
-
-    unless authorize_to_save_share_file? @share_file
-      @categories_hash = ShareFile.get_tags_hash(@share_file.owner_symbol)
-      redirect_to_with_deny_auth
-      return
-    end
+    @share_file.accessed_user = current_user
 
     if @share_file.update_attributes(update_params)
       @share_file.share_file_publications.clear
@@ -145,7 +142,10 @@ class ShareFileController < ApplicationController
   def destroy
     share_file = ShareFile.find(params[:id])
 
-    redirect_to_with_deny_auth and return unless authorize_to_save_share_file? share_file
+    unless share_file.updatable?(current_user)
+      redirect_to_with_deny_auth
+      return
+    end
 
     share_file.destroy
     flash[:notice] = _("ファイルの削除に成功しました。")
@@ -188,11 +188,13 @@ class ShareFileController < ApplicationController
     file_name =  params[:file_name]
     owner_symbol = "#{symbol_type_hash[params[:controller_name]]}:#{params[:symbol_id]}"
 
-    # ログインユーザが対象ファイルをダウンロードできるか否か判定
-    find_params = ShareFile.make_conditions(login_user_symbols, { :file_name => file_name, :owner_symbol => owner_symbol })
-    unless share_file = ShareFile.find(:first, :conditions => find_params[:conditions], :include => find_params[:include] )
-      flash[:warning] = _('指定されたファイルが存在しないか、アクセス権がありません。') # 本来は存在する場合でも、ファイルの存在自体を知らせないために、存在しない旨を表示
-      return redirect_to(:controller => 'mypage', :action => 'index')
+    unless share_file = ShareFile.find_by_file_name_and_owner_symbol(file_name, owner_symbol)
+      raise ActiveRecord::RecordNotFound
+    end
+
+    unless share_file.readable?(current_user)
+      redirect_to_with_deny_auth
+      return
     end
 
     if downloadable?(params[:authenticity_token], share_file)
@@ -217,7 +219,7 @@ class ShareFileController < ApplicationController
   def download_history_as_csv
     share_file = ShareFile.find(params[:id])
 
-    unless authorize_to_save_share_file? share_file
+    unless share_file.readable?(current_user)
       redirect_to_with_deny_auth
       return
     end
@@ -226,18 +228,19 @@ class ShareFileController < ApplicationController
     send_data(csv_text, :filename => nkf_file_name(file_name), :type => 'application/x-csv', :disposition => 'attachment')
   end
 
+  # ajax action
   def clear_download_history
     unless share_file = ShareFile.find(:first, :conditions => ["id = ?", params[:id]])
       return false
     end
 
-    unless authorize_to_save_share_file? share_file
-      render :nothing => true
+    unless share_file.updatable?(current_user)
+      render :text => _('この操作は、許可されていません。'), :status => :forbidden
       return
     end
 
     share_file.share_file_accesses.clear
-    render :nothing => true
+    render :text => _('ダウンロード履歴の削除に成功しました。')
   end
 
 private
@@ -267,13 +270,4 @@ private
     return  NKF::nkf('-Ws', file_name) if agent.include?("MSIE") and not agent.include?("Opera")
     return file_name
   end
-
-  # 権限チェック
-  # ・所有者がマイユーザ
-  # ・所有者がマイグループ  AND 作成者がマイユーザ
-  def authorize_to_save_share_file? share_file
-    session[:user_symbol] == share_file.owner_symbol ||
-      (login_user_groups.include?(share_file.owner_symbol) && (session[:user_id] == share_file.user_id))
-  end
-
 end
